Fix Calculation and Auto generation

+++ lib/models/reward.dart
@@ -1,6 +1,6 @@
 class Reward {
   final String name;
-  final int aliens;
+  final double aliens;
   final int powerupId;
 
   Reward({required this.name, this.aliens = 0, this.powerupId = 0});
