+++ lib/ui/store_page.dart
@@ -37,7 +37,7 @@ class _PowerUpShopState extends State<PowerUpShop> {
   }
 
   double _calculatePrestigeMultiplier() {
-    int totalAliens = _gameState.user!.aliens;
+    double totalAliens = _gameState.user!.aliens;
     int totalPowerUps = _gameState.powerUps.fold(0, (sum, powerUp) => sum + powerUp.purchaseCount);
     double prestigeMultiplier = 0.01;
